final-test/pyproject.toml
[project]
name = "final-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.25.1",
    "fpdf>=1.7.2",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
